警告無視

+++ Matrix.cpp
@@ -85,7 +85,7 @@ bool InverseMatrix(const Matrix& mat, Matrix& invMat)
         /* 操作（１）：k行目とmax_i行目を入れ替える */
         if (k != max_i) {
             for (int x = 0; x < 8; x++) {
-                double tmp = sweep[max_i][x];
+                float tmp = sweep[max_i][x];
                 sweep[max_i][x] = sweep[k][x];
                 sweep[k][x] = tmp;
             }
